requirements.txt
einops
jaxtyping
jax
beartype==0.15.0
black
matplotlib
ruff
scipy
trimesh
pytest
tqdm
torch
torchvision
opencv-python==4.9.0.80
scikit-image
pygame==2.6.1
PyOpenGL==3.1.0
numpy==1.24.1
Pillow
tqdm
